requirements.txt
django
python-decouple
pillow
djangorestframework
markdown       
django-filter  
drf-yasg[validation]
django-coreapi
djoser
django-mail-templated
djangorestframework-simplejwt[crypto]
black
flake8
pytest
pytest-django